Add bootstrap (#6)
Just wanted to get some UI basics in place with React Bootstrap.

I figure that this along with FontAwesome or similar icon pack will be more than enough for what I need. Probably just going to use "emoji as icon" for now if it comes up.

Also, added the ability to "sign out" of a session. This sets states back to initial so that you can "sign in" as someone else. 

This is what would happen if you lent someone a tablet to do this on.