Support Size, Number, and Mixed treatments (#23)
# Background

The treatment in this study is the kind of prompts the child is given.

There are three possible treatments:

* Questions about the size of the shapes
* Questions about the number of the shapes
* A mixture of size and number questions

The "mixed" treatment is hypothesized to yield the best results as it requires more cognition to change what should be discriminated about the stimuli.

# Work done

The "mixed" treatment necessitated a new concept of the "mode" of a given question - either number or shape.
So in the number or shape _treatment_, the mode is always the same.

However, in the mixed treatment, we generate a random mode for each page and use that.

It is a bit confusing that we always generate a random "mode" and keep it in state even when in one of the stable treatments. I would like to figure out a way to make this more clear or eliminate the random mode behavior entirely when in a stable treatment.