Adaptive difficulty (#27)
# Overview

This PR makes some decent-sized changes to the content model, grading, and presentation logic to support adaptive difficulty levels.

As the child gets on a streak of correct answers, we increase the difficulty. A streak of incorrect answers decreases the difficulty. 

For now, the difficulty levels are simply easy, medium, and hard. I have it hardcoded to require only one correct or incorrect answer to trigger a difficulty change to make testing easier.

## Content changes

I introduced a new content type: Choice. A Choice includes two stimuli, the correct answers, and a difficulty level.

Each Question now takes a list of 3 choices, rather than simply having two stimuli directly on the Question.
So the expectation is that each question will have an easy, medium, and hard choice attached. I could not put that validation at the Question level, but I did add validation to the code. This way, when we build a "broken" book, we will know it early.