Pull in data live instead of at build time. Support for a preview environment. (#33)
There are two separate issues that I am tackling at once.

### Pull in data live

Before, I was pulling in data with `npm run content:export` and dumping a `content.json`. In order to make it easier for Zach and Pat to develop their eBooks, I will just pull in published content live. So updating data will be as easy as refreshing.

### Support a preview environment

We may want the ability to preview unpublished changes. I am struggling to find the best way to handle this with Netlify. I really just want to deploy the master branch again, but with a different Contentful API key. It looks like that is not supported. So I created a branch called `unpublished` and have that set to deploy using the preview API key.