Use LongText when it is provided (#43)
This change applies for any text field: Page narrative, Question narrative, and Question prompt. (Although I expect it will not be used for question prompts).

All text in the book configured in Contentful is done using a Text content type. The text field is the title of the content type and is the default place to place text. However, if you need more than 256 characters, you can drop down into the LongText field. The app will now render the long-text field when available.