Recorded query for easy use (#45)
Responses are recorded to Firebase.
Firebase is automatically extracted to BigQuery. However, in BigQuery, there is just a data column with a string containing the entire firebase document.

This query provides all production responses with proper columns with scalar values.

This would be the query to use to generate a CSV of responses. Or, the base query for a CTE before doing analysis on top of responses.

This would be a good candidate to become a view.