Scroll to top after selecting answer (#54)
There are actually a few unrelated changes in this PR. Bad form, but I was just working quickly to knock some items out.

1. Center text on credits page
1. Get rid of counter code that came with the create-react-app template. I was keeping it out for reference at first. Then I simply forgot about it.
1. Scroll to top after picking an answer or hitting next page

That last one is to help out with an issue we are having with the zoo books. With Zoo Books, all the images have a lot of empty space at the top. Because of this, the "select" button is way at the bottom of the screen. This means that you can easily skip seeing the question feedback before moving on to the next page.

This does not fix the layout issue, but it does make sure that parents will see the feedback.